Vote to move imagemin
I have moved the imagemin task to deploy. In my workflow imagemin was taking the most time (a site with 5 images, this task takes 7.1 seconds). Which becomes annoying when testing out changes quickly. I submit this for consideration. The imagemin task is pretty self contained, not manipulating filepaths only the files contents.